Rc. file for 3.610
The MDS file contains the Menu, Dialog and Stringtable together.
Note: Both version are not complete, the only thing left to do is translate the Spec change log.
Other thing, I haven't made any revisions to the en version, so there are a lot of errors. As for the pt version, have made some revisions, but I haven't finished yet.